+++ src/GPS.RandomDataGenerator.Abstractions/IGeneratorOptions.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections.Generic;
+
+namespace GPS.RandomDataGenerator.Abstractions
+{
+    public interface IGeneratorOptions<TResult>
+    {
+        IEnumerable<TResult> Generate(Random random);
+    }
+}