+++ src/GPS.RandomDataGenerator/Generators/GivenNameGenerator.cs
@@ -8,9 +8,9 @@ using GPS.RandomDataGenerator.Abstractions;
 
 namespace GPS.RandomDataGenerator.Generators
 {
-    public class GivenNameGenerator : IDataGenerator<string>
+    public class GivenNameGenerator : IDataGenerator<string>, IResetable
     {
-        private readonly Dictionary<int, Random> _randomizers = new Dictionary<int, Random>();
+        private Dictionary<int, Random> Randomizers { get; } = new Dictionary<int, Random>();
 
         public GivenNameGenerator(IServiceProvider provider, IEnumerable<string> givenNames)
         {
@@ -21,16 +21,30 @@ namespace GPS.RandomDataGenerator.Generators
         private IServiceProvider Provider   { get; }
         private string[]         GivenNames { get; }
 
+        public IEnumerable<string> Generate(Random random, int count)
+        {
+            var seed = random.GetHashCode();
+
+            if(!Randomizers.ContainsKey(seed)) Randomizers.Add(seed, random);
+
+            return Generate(seed, count);
+        }
+
         public IEnumerable<string> Generate(int? seed, int count, params object[] options)
         {
             seed ??= DateTime.Now.Millisecond;
-            if (!_randomizers.TryGetValue(seed.Value, out var random))
+            if (!Randomizers.TryGetValue(seed.Value, out var random))
             {
                 random = new Random(seed.Value);
-                _randomizers.Add(seed.Value, random);
+                Randomizers.Add(seed.Value, random);
             }
 
             for (var i = 0; i < count; ++i) yield return GivenNames[random.Next(0, GivenNames.Length - 1)];
+        }
+ 
+        public void Reset(int seed)
+        {
+            if(Randomizers.ContainsKey(seed)) Randomizers.Remove(seed);
         }
     }
 }