+++ src/GPS.RandomDataGenerator/Options/GenerateStrings.cs
@@ -1,0 +1,8 @@
+namespace GPS.RandomDataGenerator.Options
+{
+    public enum GenerateStrings
+    {
+        SurName, GivenName, FullName,
+        EmailAddress
+    }
+}