+++ tests/GPS.RandomDataGenerator.Tests/GuidTests.cs
@@ -1,6 +1,9 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using GPS.RandomDataGenerator.Generators;
+using GPS.RandomDataGenerator.Extensions;
+using GPS.RandomDataGenerator.Options;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using Xunit.Abstractions;
@@ -25,7 +28,7 @@ namespace GPS.RandomDataGenerator.Tests
         [InlineData(0x0, 0xA)]
         [InlineData(0x1, 0xA)]
         [InlineData(0x0, 0x2710)]
-        public void GenerateGuidAddresses(int seed, int count)
+        public void GenerateGuid(int seed, int count)
         {
             var guids = Provider.GetService<GuidGenerator>()?
                 .Generate(seed, count)?
@@ -42,5 +45,55 @@ namespace GPS.RandomDataGenerator.Tests
             _testOutputHelper.WriteLine($"End Count: {count} ----------------------------");
             #endif
         }
-    }
+
+        [Theory]
+        [InlineData(0x0, 0xA)]
+        [InlineData(0x1, 0xA)]
+        [InlineData(0x0, 0x2710)]
+        public void GenerateGuidFromRandom(int seed, int count)
+        {
+            var random = new Random(seed);
+
+            var guids = random.Generate<Guid, Guid>(Provider, count).ToList();
+
+            Assert.NotNull(guids);
+            Assert.NotEmpty(guids);
+            Assert.DoesNotContain(guids, s => s == default);
+            Assert.Equal(count, guids.Count);
+
+            #if DEBUG
+            _testOutputHelper.WriteLine($"Start Count: {count} ----------------------------");
+            guids.ForEach(guid => _testOutputHelper.WriteLine($"{guid}"));
+            _testOutputHelper.WriteLine($"End Count: {count} ----------------------------");
+            #endif
+        }
+    
+        [Theory]
+        [InlineData(0x0, 0xA)]
+        [InlineData(0x1, 0xA)]
+        [InlineData(0x0, 0x2710)]
+        public void GenerateGuidNext(int seed, int count)
+        {
+            var guids = new List<Guid>();
+            var random = new Random(seed);
+
+            for(var i = 0; i < count; ++i)
+            {
+                var guid = random.Next<Guid>(Provider);
+
+                guids.Add(guid);
+            }
+
+            Assert.NotNull(guids);
+            Assert.NotEmpty(guids);
+            Assert.DoesNotContain(guids, s => s == default);
+            Assert.Equal(count, guids.Count);
+
+            #if DEBUG
+            _testOutputHelper.WriteLine($"Start Count: {count} ----------------------------");
+            guids.ForEach(guid => _testOutputHelper.WriteLine($"{guid}"));
+            _testOutputHelper.WriteLine($"End Count: {count} ----------------------------");
+            #endif
+        }
+}
 }