+++ src/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - Added Options and Extensions
+
+- Reusable Option objects:
+  - `Options`: Basic for types that don't have a range
+  - `RangeOptions`: For types that have a range
+- Extension methods on `System.Random` that can accept options or typed parameters.
+- `IResettable` allows Generators (`IDataGenerator`) to be reset.
+
 ## [1.0.3] - Packages Cleanup
 
 ### Added
@@ -26,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Define properties in a map and apply to the Simple Record.
 - Dependency Injection
 
-[Unreleased]: https://github.com/gatewayprogrammingschool/RandomDataGeneration/compare/1.0.3...HEAD
+[Unreleased]: https://github.com/gatewayprogrammingschool/RandomDataGeneration/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/gatewayprogrammingschool/RandomDataGeneration/compare/1.0.3...1.1.0
 [1.0.3]: https://github.com/gatewayprogrammingschool/RandomDataGeneration/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/gatewayprogrammingschool/RandomDataGeneration/releases/tag/1.0.2